requirements.txt
django==5.0.1
djangorestframework==3.15.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1